Move the rest of the npm FS workers back to the plugin
