+++ flow-maven-plugin/src/main/java/com/vaadin/flow/plugin/common/ClassPathIntrospector.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.server.frontend;
+package com.vaadin.flow.plugin.common;
 
 import java.io.Serializable;
 import java.lang.annotation.Annotation;
