+++ flow-maven-plugin/src/main/java/com/vaadin/flow/plugin/common/NodeUpdatePackages.java
@@ -40,8 +40,6 @@ import org.apache.commons.io.FileUtils;
 import com.vaadin.flow.component.dependency.HtmlImport;
 import com.vaadin.flow.component.dependency.JsModule;
 import com.vaadin.flow.component.dependency.NpmPackage;
-import com.vaadin.flow.server.frontend.AnnotationValuesExtractor;
-import com.vaadin.flow.server.frontend.NodeUpdater;
 
 import elemental.json.Json;
 import elemental.json.JsonObject;
