+++ flow-maven-plugin/src/main/java/com/vaadin/flow/plugin/maven/NodeUpdateImportsMojo.java
@@ -30,9 +30,9 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.plugins.annotations.ResolutionScope;
 
-import com.vaadin.flow.server.frontend.AnnotationValuesExtractor;
-import com.vaadin.flow.server.frontend.NodeUpdateImports;
-import com.vaadin.flow.server.frontend.NodeUpdater;
+import com.vaadin.flow.plugin.common.AnnotationValuesExtractor;
+import com.vaadin.flow.plugin.common.NodeUpdateImports;
+import com.vaadin.flow.plugin.common.NodeUpdater;
 
 /**
  * Goal that updates main.js file with @JsModule, @HtmlImport and @Theme
